requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Django==1.11.5
djangorestframework==3.6.4
first==2.0.1
idna==2.6
pew==0.1.26
pip-tools==1.9.0
pipenv==7.3.1
psycopg2==2.7.3.1
pythonz-bd==1.11.4
pytz==2017.2
requests==2.18.4
resumable-urlretrieve==0.1.5
semver==2.7.8
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
